Moved the PID and Odometry classes to .h files
In C++, the common practice is to have files that are meant to be included to be in other files be in .h or .hpp files. This does not need to be done, and can be reverted as I am now getting #include errors, though I have to leave for the day. If this can be implemented, though, it will probably make the code cleaner and easier to read.