Made a gyro, auton, and catapult class, as well as edited odometry to work.
All of these classes are small and were just made quickly as I wasn't  here for a day and needed to quickly get them working. You can look at them, but they are all fairly simple. A MAJOR thing that I found out though, is that the left motors on the bot are actually reversed, and the code is wrong. I will need to look into that next (basically change all the drive code) but it works well enough for now.